Split out a new DomainEventCollector from UnitOfWork for clearer responsibilities
