+++ backend-foundation/DomainModeling/DomainEvents/IDomainEventCollector.cs
@@ -1,0 +1,8 @@
+using PlatformPlatform.Foundation.DomainModeling.Entities;
+
+namespace PlatformPlatform.Foundation.DomainModeling.DomainEvents;
+
+public interface IDomainEventCollector
+{
+    IEnumerable<IAggregateRoot> GetAggregatesWithDomainEvents();
+}
