+++ backend-foundation/Tests/DomainModeling/Behaviors/PublishDomainEventsPipelineBehaviorTests.cs
@@ -3,7 +3,7 @@ using MediatR;
 using NSubstitute;
 using PlatformPlatform.Foundation.DomainModeling.Behaviors;
 using PlatformPlatform.Foundation.DomainModeling.Cqrs;
-using PlatformPlatform.Foundation.DomainModeling.Persistence;
+using PlatformPlatform.Foundation.DomainModeling.DomainEvents;
 using PlatformPlatform.Foundation.Tests.TestEntities;
 using Xunit;
 
@@ -15,10 +15,11 @@ public class PublishDomainEventsPipelineBehaviorTests
     public async Task Handle_WhenCalled_ShouldPublishDomainEvents()
     {
         // Arrange
-        var unitOfWork = Substitute.For<IUnitOfWork>();
+        var domainEventCollector = Substitute.For<IDomainEventCollector>();
         var publisher = Substitute.For<IPublisher>();
         var behavior =
-            new PublishDomainEventsPipelineBehavior<TestCommand, Result<TestAggregate>>(unitOfWork, publisher);
+            new PublishDomainEventsPipelineBehavior<TestCommand, Result<TestAggregate>>(domainEventCollector,
+                publisher);
         var request = new TestCommand();
         var cancellationToken = new CancellationToken();
         var next = Substitute.For<RequestHandlerDelegate<Result<TestAggregate>>>();
@@ -26,7 +27,7 @@ public class PublishDomainEventsPipelineBehaviorTests
 
         var testAggregate = TestAggregate.Create("TestAggregate");
         var domainEvent = testAggregate.DomainEvents.Single(); // Get the domain events that were created.
-        unitOfWork.GetAggregatesWithDomainEvents().Returns(new[] {testAggregate});
+        domainEventCollector.GetAggregatesWithDomainEvents().Returns(new[] {testAggregate});
 
         // Act
         await behavior.Handle(request, next, cancellationToken);
